pyproject.toml
[project]
name = "munra-cl"
version = "0.1.0"
description = "Lo que tienes para decir no puede esperar"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai>=0.0.42",
    "python-dotenv>=1.0.1",
    "python-fasthtml>=0.12.4",
    "sqlmodel>=0.0.8",
    "alembic>=1.12.0",
    "datasette>=0.65.1",
]
